One Corner at a time
Still, not working, but it can manage one corner at a time now that I have removed the corners left array. Its looking better, but still far from goal. To change the dot it eats, change the array indexes in the line - if (x, y) in self.corners[x:y] - from [0:1] to [3:4].